+++ src/pages/SignIn.tsx
@@ -1,7 +1,7 @@
 import { SubmitHandler, useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import { Lock, Mail } from "lucide-react";
-import useSignIn from "@/lib/auth/useSignIn";
+import useSignIn from "@/lib/auth/hooks/useSignIn";
 import { authStatusType, Layout } from "@/components/common/components/Layout";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
